Rearranged the files for Timer0/1
