+++ BasalDose/BasalDose.c
@@ -16,7 +16,7 @@ extern uint32_t StepperMotor_CurrentBolusDose;
 extern status Control_GlobalStatus;
 
 // Set and enable Timer0 for the time in between Basal doses
-void BasalDose_DoseTimingInitiate(void)
+void BasalDose_TimingInitiate(void)
 {
 	LPC_TIM0->PR = 0x02; // Pre-scalar
 	LPC_TIM0->MR0 = 1 << 27; // Match number
@@ -24,40 +24,18 @@ void BasalDose_DoseTimingInitiate(void)
 	NVIC_EnableIRQ(TIMER0_IRQn); // Enable Timer0, but don't start counting
 }
 
-void BasalDose_DoseTimingEnable(void)
+void BasalDose_TimingEnable(void)
 {
-	BasalDose_DoseDisable(); // Disable and Reset Timer1
+	StepperMotor_SpeedDisable(); // Disable and Reset Timer1
 	LPC_TIM0->TCR |= 1 << 0; // Start counting (TCR = 01)
 }
 
-void BasalDose_DoseTimingDisable(void)
+void BasalDose_TimingDisable(void)
 {
 	LPC_TIM0->TCR &=~(1 << 0); // Stop Timer Counter (TCR = 00)
 	LPC_TIM0->TCR |= 1 << 1; // Reset Timer Counter (TCR = 10)
 	LPC_TIM0->TCR &=~(1 << 1); // Stop resetting Timer Counter (TCR = 00)
 	LPC_TIM0->IR |= 1 << 0; // Reset Timer0 Interrupt
-}
-
-void BasalDose_DoseInitiate(void)
-{
-	LPC_TIM1->PR = 0x02; // Pre-scalar
-	LPC_TIM1->MR0 = 1 << 20; // Match number
-	LPC_TIM1->MCR |= 3 << 0; // Interrupt and reset timer on match (MCR = 011)
-	NVIC_EnableIRQ(TIMER1_IRQn);
-}
-
-void BasalDose_DoseEnable(void)
-{
-	BasalDose_DoseTimingDisable(); // Disable and Reset Timer0
-	LPC_TIM1->TCR |= 1 << 0; // Start counting (TCR = 01)
-}
-
-void BasalDose_DoseDisable(void)
-{
-	LPC_TIM1->TCR &=~(1 << 0); // Stop Timer Counter (TCR = 00)
-	LPC_TIM1->TCR |= 1 << 1; // Reset Timer Counter (TCR = 10)
-	LPC_TIM1->TCR &=~(1 << 1); // Stop resetting Timer Counter (TCR = 00)
-	LPC_TIM1->IR |= 1 << 1; // Reset Timer1 Interrupt
 }
 
 void TIMER0_IRQHandler(void)
@@ -77,30 +55,7 @@ void TIMER0_IRQHandler(void)
 		Control_GlobalStatus = Backward;
 		LED_On(2); // Signal that Backward/Retraction is occuring P1.31
 	}
-	BasalDose_DoseEnable();	
+	StepperMotor_SpeedEnable();	
 }
 
-void TIMER1_IRQHandler(void)
-{
-	// Switch on status defined by Timer0 and EINT3 IRQs
-	// case None defined within the StepperMotor_Step functions when adminstration is done
-	switch(Control_GlobalStatus)
-	{
-		case Basal: 
-			StepperMotor_CurrentBasalDose++; // Keep track of current dosing
-			StepperMotor_StepForward();
-			break;
-		case Bolus:
-			StepperMotor_CurrentBolusDose++; 
-			StepperMotor_StepForward();
-			break;
-		case Backward:
-			StepperMotor_StepBackward();
-			break;
-		case None:
-			BasalDose_DoseTimingEnable(); // Re-Enable Timer0
-			Control_LEDClear(); // Clear out LEDs
-			break;
-	}
-	LPC_TIM1->IR |= 1 << 0; // Clear out Timer1 registers
-}
+
